refactor: replace common Text component with new Text component in Footer, MainBanner, and StoryCard components
