+++ src/frames/Main/StoryCard.tsx
@@ -1,10 +1,8 @@
-'use client';
-
 import React from 'react';
 import styled from '@emotion/styled';
 import Link from 'next/link';
 import Image from 'next/image';
-import Text from '@krafton-soc/components/common/Text';
+import Text from '@krafton-soc/components/Text';
 import { theme } from '@krafton-soc/styles/theme';
 
 export interface StoryCardProps {
